requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.2
lxml>=5.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
